+++ lib/Extension/LanguageServerPhpstan/Configuration/PhpstanConfigSuggestor.php
@@ -5,7 +5,7 @@ namespace Phpactor\Extension\LanguageServerPhpstan\Configuration;
 use Composer\Autoload\ClassLoader;
 use Phpactor\Configurator\ChangeSuggestor;
 use Phpactor\Configurator\Model\Changes;
-use Phpactor\Configurator\Config\JsonConfig;
+use Phpactor\Configurator\Model\JsonConfig;
 use Phpactor\Extension\LanguageServerPhpstan\LanguageServerPhpstanExtension;
 
 class PhpstanConfigSuggestor implements ChangeSuggestor
